requirements.txt
MetaTrader5==5.0.5735
numpy==1.24.4
python-telegram-bot==13.15
APScheduler==3.6.3
pytz==2026.1.post1
